--- pipe_operator/python_flow/extras.py
@@ -1,67 +1,0 @@
-from typing import (
-    Callable,
-    TypeVar,
-)
-
-from typing_extensions import Concatenate, ParamSpec
-
-from pipe_operator.python_flow.base import _BasePipe
-from pipe_operator.shared.exceptions import PipeError
-from pipe_operator.shared.utils import (
-    is_async_function,
-    is_lambda,
-    is_one_arg_lambda,
-)
-
-TInput = TypeVar("TInput")
-TOutput = TypeVar("TOutput")
-FuncParams = ParamSpec("FuncParams")
-
-
-class Tap(_BasePipe[TInput, FuncParams, TInput]):
-    """
-    Pipe-able element that produces a side effect and returns the original value.
-    Useful to perform async actions or to call an object's method that changes the state
-    without returning anything.
-
-    Args:
-        f (Callable[Concatenate[TInput, FuncParams], object]): The function that will be called in the pipe.
-        args (FuncParams.args): All args (except the first) that will be passed to the function `f`.
-        kwargs (FuncParams.kwargs): All kwargs that will be passed to the function `f`.
-
-    Raises:
-        PipeError: If `f` is an async function.
-
-    Examples:
-        >>> class BasicClass
-        ...     def __init__(self, x: int) -> None:
-        ...         self.x = x
-        >>> (
-        ...     PipeStart(1)
-        ...     >> Pipe(BasicClass)
-        ...     >> Tap(lambda x: print(x))
-        ...     >> Pipe[BasicClass, [], int](lambda x: x.x + 3)
-        ...     >> Tap(lambda x: x + 100)
-        ...     >> PipeEnd()
-        ... )
-        4
-    """
-
-    def __init__(
-        self,
-        f: Callable[Concatenate[TInput, FuncParams], object],
-        *args: FuncParams.args,
-        **kwargs: FuncParams.kwargs,
-    ) -> None:
-        super().__init__(f, *args, **kwargs)  # type: ignore
-
-    def validate_f(self) -> None:
-        """f cannot be a lambda with multiple args nor an async function."""
-        if is_lambda(self.f) and not is_one_arg_lambda(self.f):
-            raise PipeError(
-                "Lambda functions with more than 1 argument are not supported."
-            )
-        if is_async_function(self.f):
-            raise PipeError(
-                "`Pipe` does not support async functions. Use `AsyncPipe` instead."
-            )
